[Refactor] AddSchedule 컴포넌트 분리 및 파일명 정리 ( #29 )
