+++ TAMINGO/Features/CalendarView/Components/AddSchedule/SchedulePlace.swift
@@ -1,0 +1,76 @@
+//
+//  SchedulePlace.swift
+//  TAMINGO
+//
+//  Created by 김도연 on 2/2/26.
+//
+
+import SwiftUI
+
+// MARK: - Place Component
+struct SchedulePlace: View {
+    // State
+    var isLoading: Bool
+    var titleInput: String
+    var placeName: String
+    
+    // Data
+    var myPlaces: [MyPlaceDTO]
+    
+    // Actions
+    var onSelectPlace: (MyPlaceDTO) -> Void
+    var onResetFavoriteRecommendation: () -> Void
+    
+    // Local State (내부적으로 관리)
+    @State private var isEditing: Bool = false
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 10) {
+            HStack(spacing: 6) {
+                ScheduleSectionHeader(title: "장소", isRequired: false)
+                AIBadge()
+            }
+            
+            if isLoading {
+                AILoadingRow(text: "장소 추론중 ...")
+            } else if titleInput.isEmpty {
+                GuideText(text: "제목 입력 시 AI가 장소를 추론합니다")
+            } else if isEditing {
+                placeListScroll
+            } else {
+                SelectedItemRow(title: placeName) {
+                    withAnimation {
+                        isEditing = true
+                        onResetFavoriteRecommendation()
+                    }
+                }
+            }
+        }
+    }
+    
+    private var placeListScroll: some View {
+        ScrollView(.horizontal, showsIndicators: false) {
+            HStack(spacing: 8) {
+                ForEach(myPlaces) { place in
+                    Button {
+                        onSelectPlace(place)
+                        withAnimation { isEditing = false }
+                    } label: {
+                        Text(place.name)
+                            .font(.regular12)
+                            .foregroundStyle(.black)
+                            .padding(.horizontal, 16)
+                            .padding(.vertical, 8)
+                            .background(
+                                RoundedRectangle(cornerRadius: 8)
+                                    .fill(Color.white)
+                                    .shadow(color: .black.opacity(0.1), radius: 3, x: 0, y: 2)
+                            )
+                    }
+                }
+            }
+            .padding(.vertical, 4)
+            .padding(.horizontal, 4)
+        }
+    }
+}
